+++ scripts/src/admin/ResponseTable/filterHeaderObjs.ts
@@ -1,16 +1,26 @@
-import {isArray, findIndex} from "lodash-es";
+import {isArray, findIndex, cloneDeep, concat} from "lodash-es";
 
-function filterHeaderObjs(headerObjs, dataOption) {
+function filterHeaderObjs(headerObjsOrig, dataOption) {
   let filtered = [];
+  let headerObjs = cloneDeep(headerObjsOrig);
   if (dataOption && dataOption.columnOrder && dataOption.columnOrder.length && isArray(dataOption.columnOrder)) {
       // Bring the columns in columnOrder to the front.
-      let columnsAtFront = (dataOption.columnOrder);
-      for (let colName of columnsAtFront.reverse()) {
+      let columnsAtFront = [];
+      let wildCardIndex = -1;
+      for (let i = 0; i < dataOption.columnOrder.length; i++) {
+          let colName = dataOption.columnOrder[i];
           let index = findIndex(headerObjs, e => e.id.toLowerCase() == colName.toLowerCase());
-          if (index != -1)
-              headerObjs.unshift(headerObjs.splice(index, 1)[0]);
+          if (index != -1) {
+            columnsAtFront.push(headerObjs.splice(index, 1)[0]);
+          }
+          else if (colName == "*") { // Wildcard operator.
+            wildCardIndex = i;
+          }
       }
-      filtered = headerObjs;
+      if (wildCardIndex != -1) { // Insert headerObjs into wild card index.
+        columnsAtFront.splice(wildCardIndex, 0, ...headerObjs);
+      }
+      filtered = columnsAtFront;
       // filtered = headerObjs.sort((a, b) => ~findIndex(dataOption.columnOrder, header => b.id == header));
       // filtered = intersectionWith(headerObjs, dataOption.columnOrder, (header, order) => header.id == order);
   }