+++ scripts/src/admin/ResponseTable/ResponseDetail.d.ts
@@ -1,0 +1,11 @@
+interface IResponseDetailProps {
+  checkInMode: boolean,
+  data: any,
+  formId: string,
+  responseId: string,
+  dataOptions: IDataOptions
+}
+
+interface IResponseDetailState {
+  data: any
+}