+++ scripts/src/admin/ResponseTable/ResponseDetail.tsx
@@ -1,3 +1,4 @@
+/// <reference path="./ResponseDetail.d.ts"/>
 import ReactJson from 'react-json-view';
 import * as React from 'react';
 import { API } from "aws-amplify";
@@ -5,7 +6,7 @@ import ReactTable from "react-table";
 import dataLoadingView from "../util/DataLoadingView";
 import {get, set} from "lodash-es";
 
-class ResponseDetail extends React.Component<{data: any, formId: string, responseId: string, dataOptions: IDataOptions}, any> {
+class ResponseDetail extends React.Component<IResponseDetailProps, IResponseDetailState> {
     constructor(props:any) {
         super(props);
         this.state = {
@@ -14,7 +15,7 @@ class ResponseDetail extends React.Component<{data: any, formId: string, respons
     }
     changeCheckIn(row, checked) {
         console.log(checked);
-        API.post("CFF", `forms/${this.props.formId}/responses/${this.props.responseId}/edit`, {
+        API.post("CFF", `forms/${this.props.formId}/responses/${this.props.responseId}/checkin`, {
             "body":
             {
                 "path": `value.${row.cff_accessor}.check_in`, // participants[0].check_in
@@ -61,19 +62,15 @@ class ResponseDetail extends React.Component<{data: any, formId: string, respons
             Header: "Check in",
             accessor: "check_in",
             Cell: ({original}) => <input type="checkbox" checked={original.check_in} onChange={e => this.changeCheckIn(original, e.target.checked)} value="" id={`defaultCheck_${this.props.responseId}`} />
-        },
-        // {
-        //     Header: "Accessor",
-        //     accessor: "cff_accessor"
-        // }
+        }
     ];
     let i = 0;
     let tableData = this.state.data.value.participants.map(e => Object.assign({"cff_accessor": `participants.${i++}`}, e));
-    let showOmrunTable = get(this.props.dataOptions, "mainTable.omrunCheckin") == true;
+    let showOmrunTable = get(this.props.dataOptions, "checkinTable.omrunCheckin") == true;
         return (
             <div className="container-fluid" key={this.props.responseId}>
                 <div className="row">
-                {showOmrunTable && <div className="card col-12 col-sm-4">
+                {this.props.checkInMode && showOmrunTable && <div className="card col-12">
                     <div className="card-body">
                         <h5 className="card-title">Details</h5>
                             <ReactTable
@@ -89,15 +86,9 @@ class ResponseDetail extends React.Component<{data: any, formId: string, respons
                                     }
                                   }}
                             />
-                            {/* <div className="form-check">
-                                <input className="form-check-input" type="checkbox" value="" id={`defaultCheck_${this.props.responseId}`} />
-                                <label className="form-check-label" htmlFor={`defaultCheck_${this.props.responseId}`} >
-                                    Picked up bib? 
-                                </label>
-                            </div> */}
                     </div>
                 </div>}
-                <div className="card col-12 col-sm-4">
+                {!this.props.checkInMode && <div className="card col-12">
                     <div className="card-body">
                         <h5 className="card-title">Inspector</h5>
                         <ReactJson src={this.state.data}
@@ -110,7 +101,7 @@ class ResponseDetail extends React.Component<{data: any, formId: string, respons
                             style={{ "fontFamily": "Arial, sans-serif", "marginLeft": "30px" }}
                         />
                     </div>
-                </div>
+                </div>}
                 </div>
             </div>
         );