+++ CHANGELOG.md
@@ -42,6 +42,8 @@ Todo:
 ["name.first", "name.last", "email"] -- only 3 columns
 ["name.first", "name.last", "email", "*"] -- 3 columns + more.
 
+- Dev: Use TDD with Jest
+
 ## 1.3.18 (4/28/18)
 - Allow paypal cancel url to reflect iframe's parent, not cff url.
 - Bug fix: Form checkbox and round off are now required when specified.