+++ CHANGELOG.md
@@ -37,12 +37,13 @@ Todo:
 - Let minItems be specified dynamically based on other values.
 - Bug: if paid amount > actual amount, it says amount paid = actual amount, not the actual paid amount!!!
 
-## 1.3.19 (tbd)
+## 1.3.19 (4/28/18) (backend 1.1.5)
 - dataOptions.columnOrder by default includes *only those columns*, to include the rest of the columns, use a wildcard:
 ["name.first", "name.last", "email"] -- only 3 columns
 ["name.first", "name.last", "email", "*"] -- 3 columns + more.
 - Specific page for check in, with form permission Responses_CheckIn
 - Don't show entire response in check in page.
+- Show "None" in check in detail table if empty value.
 - Dev: Use TDD with Jest
 
 ## 1.3.18 (4/28/18)